Initial commit - existing scales.py
This is the state that was reached before I started using version
control. We have a single file that can serve as a module or a script,
and but the script part needs to be edited manually in order to select
one of the variants.

The module has classes that allow to:
- print the notes names for each major and minor (chromatic) scale
- compute possible fingerings of each scale and sort them so that the
standard fingering comes first
- print out some details about why fingerings were sorted that way

All names are in French, but the English and German systems are
supported as well by editing the definition of a few variables.